Major modifications done
