+++ include/video_stream_opencv/video_stream_opencv.hpp
@@ -1,0 +1,7 @@
+#ifndef _VIDEO_STREAM_OPENCV_VIDEO_STREAM_OPENCV_HPP_
+#define _VIDEO_STREAM_OPENCV_VIDEO_STREAM_OPENCV_HPP_
+
+#include <video_stream_opencv/image_capture.hpp>
+#include <video_stream_opencv/utility.hpp>
+
+#endif  // ifndef _VIDEO_STREAM_OPENCV_VIDEO_STREAM_OPENCV_HPP_
