+++ package.xml
@@ -23,6 +23,7 @@
   <build_depend>image_transport</build_depend>
   <build_depend>camera_info_manager</build_depend>
   <build_depend>roscpp</build_depend>
+  <build_depend>roslib</build_depend>
   <build_depend>rospy</build_depend>
   <build_depend>sensor_msgs</build_depend>
 
@@ -30,6 +31,7 @@
   <run_depend>image_transport</run_depend>
   <run_depend>camera_info_manager</run_depend>
   <run_depend>roscpp</run_depend>
+  <run_depend>roslib</run_depend>
   <run_depend>rospy</run_depend>
   <run_depend>sensor_msgs</run_depend>
